When casting to/from a pointer to an integral type (a bad practice, if you ask me), you need to cast via intptr_t for portability.

+++ src/racoon/isakmp.c
@@ -1152,7 +1152,7 @@ isakmp_ph1begin_r(msg, remote, local, etype)
 	/* check if this etype is allowed */
 	memset(&rmsel, 0, sizeof(rmsel));
 	rmsel.remote = remote;
-	if (enumrmconf(&rmsel, check_etypeok, (void *) (unsigned) etype) == 0) {
+	if (enumrmconf(&rmsel, check_etypeok, (void *) (intptr_t) etype) == 0) {
 		plog(LLV_ERROR, LOCATION, remote,
 		     "exchange %s not allowed in any applicable rmconf.\n",
 		     s_isakmp_etype(etype));
